Allow arbitrary extra options in runMOFA
For example, you could pass seed=12345 as an additional argument to
runMOFA, and it would add "--seed 12345" to the command line. This
does not allow manually overriding any of the command-line arguments
that runMOFA sets based on the values in the mofa object itself (e.g.
you cannot add inFiles as an extra option).

Also adds a verbose option that echoes the command to be run as well
as adding "--verbose" to the command itself.

Lastly, this commit fixes a probable bug that could unintentionally
duplicate all the arguments before to "--scale_covariates" if there
are multiple covariates in the model.